Move bypass muxes from RF to EX
Also moves everything after it (like unpacking and operand muxes) to EX.

I originally thought they were in EX, because I was assuming that the
"shift muxes" on the r4300i datasheet were showing pipeline registers, and
the other pipeline registers were just missing from the diagram.
Also, EX seemed to be very bloated already.

But it wasn't possible to implement LDI interlocks as the documentation
described while "bypass muxes" were in RF, as the instruction waiting
for load data had already moved to EX. This made me suspect I was wrong.

But then I took a closer look at figure 2 in US5450607, and it explictly
shows where the pipeline registers are, and the bypass muxes (and unpacking)
are after RF.